Rename: 추상화를 위한 패키지명 수정
커뮤니티 기능 추상화를 위해 커뮤니티의 패키지명을 변경하였다.

Related to: #37
