+++ src/main/java/pulleydoreurae/careerquestbackend/basiccommunity/exception/CommunityException.java
