+++ src/main/java/pulleydoreurae/careerquestbackend/basiccommunity/repository/PostLikeRepository.java
